LayoutTests/storage/indexeddb/versionchangerequest-activedomobject-expected.txt

Ensure that IDBVersionChangeRequest objects are not GC'd if they have pending events

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;
PASS Boolean(indexedDB && IDBCursor && IDBDatabase && IDBDatabaseError && IDBDatabaseException && IDBFactory && IDBIndex && IDBKeyRange && IDBObjectStore && IDBRequest && IDBTransaction) is true


window.dbname = 'versionchangerequest-activedomobject'

testDeleteDatabase():
indexedDB.deleteDatabase(window.dbname)
window.gc()
PASS deleteDatabase's IDBVersionChangeRequest.onsuccess event fired

testSetVersion():
openRequest = indexedDB.open(window.dbname)
db = openRequest.result
db.setVersion('1')
window.gc()
PASS setVersion's IDBVersionChangeRequest.onsuccess event fired
PASS successfullyParsed is true

TEST COMPLETE
